@@ -25,7 +25,9 @@
 	},
 	"typesVersions": {
 		">4.0": {
-			"providers": ["./dist/providers/index.d.ts"]
+			"providers": [
+				"./dist/providers/index.d.ts"
+			]
 		}
 	},
 	"files": [
@@ -42,9 +44,9 @@
 		"@skeletonlabs/skeleton": "^2.8.0",
 		"@skeletonlabs/tw-plugin": "^0.3.1",
 		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
+		"@sveltejs/kit": "^2.5.27",
 		"@sveltejs/package": "^2.2.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/eslint": "8.56.0",
 		"@types/node": "^20.11.19",
 		"@typescript-eslint/eslint-plugin": "^6.0.0",
@@ -52,19 +54,19 @@
 		"autoprefixer": "^10.4.16",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.45.1",
 		"postcss": "^8.4.32",
 		"postcss-load-config": "^5.0.2",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.5.9",
 		"publint": "^0.1.9",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.3.6",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.11",
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4",
 		"@floating-ui/dom": "^1.6.3"
 	},
 	"svelte": "./dist/index.js",
