@@ -4,25 +4,39 @@
 	import type { ShareDrawerData } from './ShareDrawerSettings.js';
 	import type { CssClasses } from '@skeletonlabs/skeleton';
 
-    export let provider:ShareProvider;
-    export let shareData:ShareDrawerData;
-    export let style:{
-        background?: CssClasses,
-        shadow?: CssClasses,
-        rounded?: CssClasses,
-        overflow?: CssClasses,
-        text?: CssClasses
-    }
+	interface Props {
+		provider: ShareProvider;
+		shareData: ShareDrawerData;
+		style: {
+			background?: CssClasses;
+			shadow?: CssClasses;
+			rounded?: CssClasses;
+			overflow?: CssClasses;
+			text?: CssClasses;
+			onclick?: () => void;
+		};
+		[key: string]: any;
+	}
 
-    $: classes = `${style.background} ${style.shadow} ${style.rounded} ${style.overflow} ${$$props.class ?? ''}`
+	let { onclick, ...props }: Props = $props();
+
+	let classes = $derived(
+		`${props.style.background} ${props.style.shadow} ${props.style.rounded} ${props.style.overflow} ${props.class ?? ''}`
+	);
 </script>
-<a href={provider.getShareUrl(shareData)} on:click target="_blank" class="btn p-0 flex flex-col justify-center items-center snap-center min-w-[72px]">
-	<div class="{classes}" class:p-2={!provider.image} >
-        {#if provider.image}
-            <img src={provider.image} alt={provider.name} class="h-[72px] w-[72px]" >
-        {:else}
-            <Icon icon={provider.icon} height={56} />
-        {/if}
+
+<a
+	href={props.provider.getShareUrl(props.shareData)}
+	{onclick}
+	target="_blank"
+	class="btn flex min-w-[72px] snap-center flex-col items-center justify-center p-0"
+>
+	<div class={classes} class:p-2={!props.provider.image}>
+		{#if props.provider.image}
+			<img src={props.provider.image} alt={props.provider.name} class="h-[72px] w-[72px]" />
+		{:else}
+			<Icon icon={props.provider.icon} height={56} />
+		{/if}
 	</div>
-    <p class={style.text}>{provider.name}</p>
+	<p class={props.style.text}>{props.provider.name}</p>
 </a>
