@@ -1,7 +1,9 @@
-<div class="bg-surface-100-800-token p-2 flex justify-between">
-    <p>Made by Philip Flyvholm</p>
-    <div class="flex gap-2">
-        <a href="https://github.com/PhilipFlyvholm/skeleton-share" target="_blank" class="anchor">Github</a>
-        <a href="https://twitter.com/PhilipFlyvholm" target="_blank" class="anchor">X</a>
-    </div>
+<div class="bg-surface-100-800-token flex justify-between p-2">
+	<p>Made by Philip Flyvholm</p>
+	<div class="flex gap-2">
+		<a href="https://github.com/PhilipFlyvholm/skeleton-share" target="_blank" class="anchor"
+			>Github</a
+		>
+		<a href="https://twitter.com/PhilipFlyvholm" target="_blank" class="anchor">X</a>
+	</div>
 </div>
